Main functionality is implemented but some core features are missing

+++ shared_company/src/lib.rs
@@ -66,15 +66,16 @@ impl SharedCompany {
         let percentage_of_all_shares = shares.amount() / self.share_counter;
         // Decreases the counter
         self.share_counter -= shares.amount();
-        //ToDo Burns the shares
+        //ToDoBurns the shares
         //ToDO Burn the voting_token
+
         // returns the same percentage of the company xrd
         self.company_radix.take(self.company_radix.amount() * percentage_of_all_shares)
     }
 
     // A proposal that if it is accepted sends funds away from the company
     pub fn make_proposal(&self, destination_adress: String,reason: String,
-        admin_adress: String, end_epoch: u32,){
+        admin_adress: String, end_epoch: u64,){
         //ToDo change this to smt variable, but fails method call for some reason
             let cost = self.company_radix.take(5);
         let proposal =  Proposal::new(cost, destination_adress, reason, admin_adress, end_epoch, self.share_counter / 2 + 1, self.company_voting_token.resource_def());
